refactor(db): extract SearchStore into src/server/stores/search-store.ts (#144)
Moves FTS5 full-text search + LIKE-fallback + index sync into the third
and final focused store. Behaviour is bit-for-bit identical — pinned by
the db-search.test.ts safety net merged in #161.

Surface:
- SearchStore.searchStashes(query, options) — BM25 ranking, snippet
  detection via U+E000/U+E001 sentinel, tag/archived filters, paging.
- SearchStore.buildFtsQuery(input) — token sanitizer (strips FTS5
  operators, caps at 50 tokens / 2000 chars, 2+ char prefix scans).
- SearchStore.syncIndex / removeIndex / rebuildIndex — FTS5 virtual
  table lifecycle, called from ClawStashDB create/update/delete/import.

ClawStashDB keeps a one-line searchStashes delegator + thin wrappers
for the FTS sync calls (syncFtsIndex / removeFtsIndex / rebuildFtsIndex
are still used by createStash, updateStash, deleteStash, importAllData).
LIKE fallback callback rewires through listStashes which still lives on
ClawStashDB.

Numbers: src/server/db.ts 1400 -> 1161 LoC (-239). With this PR db.ts
shrinks from 1549 (start of session) to 1161 — total -388 LoC across
the three PRs of #144. 107 tests green, tsc clean, build green.
